requirements.txt
Bottleneck==1.3.2
certifi==2021.10.8
joblib==1.1.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numexpr==2.7.3
numpy==1.21.2
pandas==1.3.4
pip==21.0.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
scikit-learn==1.0.1
scipy==1.7.1
setuptools==58.0.4
six==1.16.0
threadpoolctl==2.2.0
wheel==0.37.0
wincertstore==0.2
